+++ swap/src/asb/rpc/server.rs
@@ -253,6 +253,7 @@ impl AsbApiServer for RpcImpl {
                 xmr_amount: state3.xmr.as_pico(),
                 exchange_rate,
                 btc_redeem_fee: state3.tx_redeem_fee,
+                btc_redeem_address: state3.redeem_address().to_string(),
                 btc_redeem_txid: state3.tx_redeem().txid().to_string(),
                 btc_punish_txid: state3.tx_punish().txid().to_string(),
                 peer_id: peer_id.to_string(),